Add ProductService interface, implementation of ProductService and JUnit test for ProductServiceImpl
Project: clothingandfootwearstore

Add ProductService interface, implementation of ProductService and JUnit test for ProductServiceImpl

+++ src/main/java/com/mateusz/model/Boots.java
@@ -4,7 +4,7 @@ public class Boots extends Product {
     private int size;
     private boolean isNatualSkin;
 
-    public Boots(int id, String productName, int price, int weight, String color, int productCount, int size, boolean isNatualSkin) {
+    public Boots(int id, String productName, float price, float weight, String color, int productCount, int size, boolean isNatualSkin) {
         super(id, productName, price, weight, color, productCount);
         this.size = size;
         this.isNatualSkin = isNatualSkin;
