+++ src/main/java/com/mateusz/model/Product.java
@@ -3,12 +3,12 @@ package com.mateusz.model;
 public class Product {
     private int id;
     private String productName;
-    private int price;
-    private int weight;
+    private float price;
+    private float weight;
     private String color;
     private int productCount;
 
-    public Product(int id, String productName, int price, int weight, String color, int productCount) {
+    public Product(int id, String productName, float price, float weight, String color, int productCount) {
         this.id = id;
         this.productName = productName;
         this.price = price;
@@ -26,11 +26,11 @@ public class Product {
         return productName;
     }
 
-    public int getPrice() {
+    public float getPrice() {
         return price;
     }
 
-    public int getWeight() {
+    public float getWeight() {
         return weight;
     }
 
@@ -43,7 +43,7 @@ public class Product {
     }
 
     // Setters
-    public void setPrice(int price) {
+    public void setPrice(float price) {
         this.price = price;
     }
